Add Go modules support
This change introduces one global Go module for Boruta. It would be
better to split it to more modules - for main boruta package, binaries
and HTTP client. Unfortunately with current state of Go modules it can't
be done in simple and elegant way. It is possible that in the future
split may occur if Go modules starts handling such situations.

Signed-off-by: Maciej Wereski <maciey@members.fsf.org>
Change-Id: I3d33f9963a33c9984c5a138ab53584bd05159da0